requirements.txt
alembic==1.6.5
flask==1.0.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
sqlalchemy==1.3.4
flask_sqlalchemy==2.5.0
sqlalchemy_utils==0.38.2
psycopg2-binary==2.9.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
python-dotenv==0.20.0
Mako==1.1.4
MarkupSafe==2.0.1
python-jose-cryptodome==1.3.2
Click==7.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.12.0
requests==2.27.1
Werkzeug==2.0.1
Jinja2==3.0.1
